Added evaluation methods as well as tests

+++ C++/src/sparsification/eta_k_core.hpp
@@ -12,13 +12,12 @@ namespace uu {
 namespace net {
 
 /**
- * (eta,k)-core
  * This function returns the (eta_k)-core value of all nodes in a probabilistic graph introduced in paper "Core Decomposition of Uncertain Graphs"
  * @param original_graph input probabilistic graph
  * @param eta_threshold eta threshold
- * @return the (eta,k)-core of a proabilistic graph
+ * @return the (eta,k)-core of a probabilistic graph
   **/
-std::vector<double>
+std::vector<int>
 eta_k_core(
     ProbabilisticNetwork * original_graph,
     double eta_threshold
